Further proof attempts
Added lemma stating an induced subgraph is a valid game and one to show that there is always a cycle in that subgraph,
Further work on closedness proof for winning region strategies. It may be necessary to include closedness in the definition rather than proving it.
Added definition for the nodes in an induced subgraph and proved some lemmata about this definition.