Work on case empty B
Combined strategies for A and winning region for alpha.
Added random choice for v because it is not included in the strategy for A.
Showed this strategy wins if it does not go through A at all.
Showing it wins if it goes through A still needs to be done.