Van Dijk's lemma 2 separated, minor cleanup, WIP
Work is in progress on proving that the discovered tangles are new. This may be scrapped if there is not enough time.

The non-emptiness of Y after search has been moved outside the invariant, because it can be empty if the initial R is empty, which cannot easily be expressed in the invariant.